feat(scripts): preview clip generator — mp4-only, @offset naming, ADR 0003
Rename generate-preview.js to generate-previews.js so the existing
generate:previews npm script actually resolves. Clips are cut to ~3s,
silent, 640px H.264 with +faststart; the name@8.mp4 convention encodes
the start offset. ADR 0003 records why clips are generated locally and
committed rather than built in CI (no ffmpeg on Netlify).
